refactor(rlp-protocol): move the Web3 transaction RLP model out of bcos-rpc
The transaction admission layer has to decode a signed Web3 envelope, but the
only decoder lives in bcos-rpc, and rpc links txpool -- so anything txpool links
cannot reach it without a cycle. Web3Transaction is the Ethereum-transaction
counterpart of EthBlockHeader, which already sits in bcos-rlp-protocol as a
"pure RLP bridge", so that is where it belongs.

Moved verbatim (git records all four as renames):

  Web3Transaction.{h,cpp}  bcos-rpc/web3jsonrpc/model -> bcos-rlp-protocol
  Web3TxHandler.{h,cpp}    bcos-rpc/web3jsonrpc/model -> bcos-rlp-protocol

Web3TxHandler was already free of tars and rpc dependencies, so it moved as-is.
Web3Transaction had exactly one member touching tars, takeToTarsTransaction().
Rather than rewrite it into a free function and change every call site, its
declaration stays with the class over a forward-declared bcostars::Transaction,
and the body moves verbatim to bcos-tars-protocol/protocol/Web3TarsBridge.cpp.
protocol-tars already links rlp-protocol, and every caller necessarily links
protocol-tars -- that is what gives them a bcostars::Transaction to receive. So
`web3Tx.takeToTarsTransaction()` still compiles unchanged everywhere, and
bcos-rlp-protocol gains no tars dependency: it carries a forward declaration and
nothing more.

Consumers only change their include path (7 files). Two stale path comments and
the `#include <bcos-rpc/jsonrpc/Common.h>` that the moved .cpp no longer needs
are updated/dropped. No behaviour change.

Deviations from the plan, both deliberate:
  - Filenames keep the Web3* spelling instead of being renamed to EthTransaction.
    Renaming on top of moving would defeat git's rename detection and make the
    diff unreviewable, and Web3TxEnvelope.{h,cpp} already uses this spelling in
    the same directory.
  - Web3TypeTest.cpp stays in bcos-rpc/test. It is a BOOST_FIXTURE_TEST_SUITE over
    RPCFixture, so moving it would drag the whole rpc stack into the rlp-protocol
    test target; it already exercises the moved code through the new include path.

Verified:
  cmake --build build --target rpc protocol-tars rlp-protocol executor
        test-bcos-rpc test-bcos-rlp-protocol            -> all built clean
  ./build/bcos-rpc/test/test-bcos-rpc                   -> No errors detected
  ./build/bcos-rlp-protocol/test/test-bcos-rlp-protocol -> No errors detected

Each moved/new .cpp was additionally compiled standalone with its target's real
flags (bypassing the unity batching that can hide a missing include). That check
found a missing bcos-framework/protocol/Transaction.h in Web3TarsBridge.cpp,
fixed here. It also reports a pre-existing failure in Web3TxHandler.cpp, which
does not compile standalone on the base commit either (an ADL issue in
bcos-codec's lengthOfItems); unchanged by this move and left alone.
