+++ bcos-executor/src/Web3AccessListResolver.cpp
@@ -3,7 +3,7 @@
 #include "bcos-codec/rlp/RLPDecode.h"
 #include "bcos-framework/protocol/Transaction.h"
 #include "bcos-framework/protocol/Web3AccessList.h"
-#include "bcos-rpc/web3jsonrpc/model/Web3Transaction.h"
+#include "bcos-rlp-protocol/Web3Transaction.h"
 #include "bcos-utilities/BoostLog.h"
 #include "bcos-utilities/Common.h"
