+++ bcos-rlp-protocol/bcos-rlp-protocol/Web3TxHandler.cpp
@@ -1,6 +1,6 @@
-// bcos-rpc/bcos-rpc/web3jsonrpc/model/Web3TxHandler.cpp
+// bcos-rlp-protocol/bcos-rlp-protocol/Web3TxHandler.cpp
 #include "Web3TxHandler.h"
-#include "Web3Transaction.h"
+#include "bcos-rlp-protocol/Web3Transaction.h"
 #include <bcos-utilities/DataConvertUtility.h>
 #include <bcos-utilities/Log.h>
 #include <cstddef>  // std::ptrdiff_t (ListEnd pointer arithmetic)
